Input Mapping, Defaults, Velocity, CarryingState
Rather than adding force, we're updating the velocity directly now. When carrying It velocity is reduced by 25%. Only assign reference of It if It is null. Moved defaults to an inner class and exposed key names as public fields, allowing input remapping. Space is the default key for jump, but the scene has it remapped to Z for now, until we make the proper adjustment.